+++ ComponentTester-1.50m/Files/functions.h
@@ -230,7 +230,7 @@
 #ifdef INA226_CURRENT_SENSOR
 
 extern uint8_t INA226_setup();
-extern int32_t INA226_getLoadVoltage_mV(void);
+extern uint16_t INA226_getLoadVoltage_mV(void);
 extern int32_t INA226_getCurrent_uA(void);
 
 #endif
